Change undo/redo and (de-)serialization behavior
Undo & redo works now as intended. Intermediate values are now not
stored in the undo/redo history (e.g., dragging a slider). When
dragging around objects it is now also properly represented in the
history.

The scene editor now also holds a "working copy" of the serialized
scene. This is used for drawing the ui for components, so they are
not deserialized every frame but only once after editing a value
is finished. This makes editing less "flickery" when the serialize
+ deserialize process does lead to slightly different values.